% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collectTrioScores.R
\name{collectTrioScores}
\alias{collectTrioScores}
\title{Collect all the scores for possible parents for an offspring.}
\usage{
collectTrioScores(kidId, damIds, sireIds, animalAlleles)
}
\arguments{
\item{kidId}{Character vector with the \code{refId} of the offspring.}

\item{damIds}{Character vector of \code{refId} values corresponding to
possible dams for the offspring.}

\item{sireIds}{Character vector of \code{refId} values corresponding to
possible sires for the offspring.}

\item{animalAlleles}{List with all allele information for all animals.}
}
\description{
This takes a single offspring and a list of potential dams and
a list of potential dams and calculates the number of loci with
alleles \code{missingLoci} for each sire and dam, the number of
\code{invalidLoci}
loci for the offspring. Note: \code{invalidLoci} loci for the offspring are
defined the same as \code{missingLoci} loci for potential parents; they are
loci with no observed alleles.
}
