% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScores.R
\name{computescores}
\alias{computescores}
\title{Tests to see if the potential trios file has the appropriate header.}
\usage{
computescores(parent, kid)
}
\arguments{
\item{parent}{Object of type \code{animal} representing one potential parent.}

\item{kid}{A Object of type \code{animal} representing the offspring.}
}
\description{
Tests to see if the potential trios file has the appropriate header.
}
