% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countDescrepant.R
\name{countDiscrepant}
\alias{countDiscrepant}
\title{Counts the number of loci with missing data}
\usage{
countDiscrepant(parent, kid)
}
\arguments{
\item{parent}{List of parent alleles with character vector of
haplotype values.}

\item{kid}{List of offspring alleles with character vector of
haplotype values.}
}
\value{
An integer value equal to the number of loci missing data
}
\description{
Counts the number of loci with missing data
}
