% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countMissing.R
\name{countMissing}
\alias{countMissing}
\title{Counts the number of loci with missing data}
\usage{
countMissing(alleles)
}
\arguments{
\item{alleles}{List of alleles with character vector of haplotype values.}
}
\value{
An integer value equal to the number of loci missing data
}
\description{
Counts the number of loci with missing data
}
