% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeReport.R
\name{createParentFindrReport}
\alias{createParentFindrReport}
\title{Create results report for parentfindr}
\usage{
createParentFindrReport()
}
\description{
Exclusion Parentage Report: Wed May 02 16:33:23 CDT 2012
}
