% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/excludeOffspring.R
\name{excludeOffspring}
\alias{excludeOffspring}
\title{Exclude offspring by ID.}
\usage{
excludeOffspring(offspring, scoresDf)
}
\arguments{
\item{offspring}{character vector with list of offspring IDs}

\item{scoresDf}{dataframe from scores object}
}
\description{
Removes offspring from \code{scoresDf} that are listed in \code{offspring}.
}
