% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnimalAlleleFileHeaders.R
\name{getAnimalAlleleFileHeaders}
\alias{getAnimalAlleleFileHeaders}
\title{Gets character vector of the approved file headers before loci list for a
animal allele file.}
\usage{
getAnimalAlleleFileHeaders()
}
\description{
Gets character vector of the approved file headers before loci list for a
animal allele file.
}
