% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnimalAlleles.R
\name{getAnimalAlleles}
\alias{getAnimalAlleles}
\title{Tests to see if the potential trios file has the appropriate header.}
\usage{
getAnimalAlleles(lines, dateType)
}
\arguments{
\item{lines}{Character vector of length one plus the number of animals where
the additional line is the header}

\item{dateType}{A character vector of length one used to guide
interpretation of date fields. Either "YYYYMMDD" or "mm/dd/YYYY"}
}
\description{
Tests to see if the potential trios file has the appropriate header.
}
