% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAnimalAllelesDfFromExcel.R
\name{getAnimalAllelesDfFromExcel}
\alias{getAnimalAllelesDfFromExcel}
\title{Extract animal allele data from the Excel formatted animal allele file.}
\usage{
getAnimalAllelesDfFromExcel(alleleFile, dateType)
}
\arguments{
\item{alleleFile}{Character vector of length one having the path of the
animal alleles file.}

\item{dateType}{A character vector of length one used to guide
interpretation of date fields. Either "YYYYMMDD" or "mm/dd/YYYY"}
}
\description{
Extract animal allele data from the Excel formatted animal allele file.
}
