% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstParents.R
\name{getFirstParents}
\alias{getFirstParents}
\title{Get first parent IDs}
\usage{
getFirstParents(id, firstParentType, scores)
}
\arguments{
\item{id}{offspring ID}

\item{firstParentType}{character vector of length one with one of \emph{dams}
or \emph{sires}.}

\item{scores}{list object with a comparison information for all trio
combinations.}
}
\description{
Get first parent IDs
}
