% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOffspringNoDam.R
\name{getOffspringNoDam}
\alias{getOffspringNoDam}
\title{Get offspring with no dams in trios object}
\usage{
getOffspringNoDam(trios)
}
\arguments{
\item{trios}{List containing a list for each offspring. Each offspring list
contains a named character vector of length one (\code{offspring}) containing
the offspring ID, a character vector of potential dams (\code{dams}),
and a named character vector of potential sires (\code{sires}).}
}
\description{
Get offspring with no dams in trios object
}
