% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOffspringOverMaxMissing.R
\name{getOffspringOverMoxMissing}
\alias{getOffspringOverMoxMissing}
\title{Get offspring with greater than maximum missing loci}
\usage{
getOffspringOverMoxMissing(scoresDf, thMissing)
}
\arguments{
\item{scoresDf}{dataframe with scores from \emph{scores} object}

\item{thMissing}{threshold for maximum missing loci}
}
\description{
Get offspring with greater than maximum missing loci
}
