% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getParentfindrConfigFileName.R
\name{getParentfindrConfigFileName}
\alias{getParentfindrConfigFileName}
\title{getParentfindrConfigFileName returns the configuration file name appropriate
for the system.}
\usage{
getParentfindrConfigFileName(sysInfo)
}
\arguments{
\item{sysInfo}{object returned by Sys.info()}
}
\value{
Character vector with expected configuration file
}
\description{
getParentfindrConfigFileName returns the configuration file name appropriate
for the system.
}
\examples{
\donttest{
library(parentfindr)
sysInfo <- Sys.info()
config <- getParentfindrConfigFileName(sysInfo)
}
}
