% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPotentialTriosFileHeaders.R
\name{getPotentialTriosFileHeaders}
\alias{getPotentialTriosFileHeaders}
\title{Gets character vector of the approved file headers for a potential trios
file.}
\usage{
getPotentialTriosFileHeaders()
}
\description{
Gets character vector of the approved file headers for a potential trios
file.
}
