% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSecondParents.R
\name{getSecondParents}
\alias{getSecondParents}
\title{Get second parent IDs}
\usage{
getSecondParents(id, firstParent, firstParentType, scores)
}
\arguments{
\item{id}{offspring ID}

\item{firstParent}{first parent ID}

\item{firstParentType}{character vector of length one with one of \emph{dams}
or \emph{sires}.}

\item{scores}{list object with a comparison information for all trio
combinations.}
}
\description{
Get second parent IDs
}
