% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSecondParentScores.R
\name{getSecondParentScores}
\alias{getSecondParentScores}
\title{Get second parent score from scores object}
\usage{
getSecondParentScores(
  id,
  firstParent,
  secondParent,
  firstParentType,
  columnNames,
  scores
)
}
\arguments{
\item{id}{progeny ID}

\item{firstParent}{first parent ID}

\item{secondParent}{second parent ID}

\item{firstParentType}{character vector of length one with one of \emph{dams}
or \emph{sires}.}

\item{columnNames}{character vector of columns to include from \code{scores}
in dataframe}

\item{scores}{list object with a comparison information for all trio
combinations.}
}
\description{
Get second parent score from scores object
}
