% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSublist.R
\name{getSublist}
\alias{getSublist}
\title{Get the names in a named sublist nested at any depth within the list
provided.}
\usage{
getSublist(x, nodeName)
}
\arguments{
\item{x}{The list to be searched}

\item{nodeName}{The node name within the list for which the names are sought}
}
\value{
A character vector of names found.
}
\description{
Get the names in a named sublist nested at any depth within the list
provided.
}
