% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrios.R
\name{getTrios}
\alias{getTrios}
\title{Tests to see if the potential trios file has the appropriate header.}
\usage{
getTrios(triosFile)
}
\arguments{
\item{triosFile}{Character vector of length one having the path of the
trios file.}
}
\description{
Tests to see if the potential trios file has the appropriate header.
}
