% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTriosDfFromExcel.R
\name{getTriosDfFromExcel}
\alias{getTriosDfFromExcel}
\title{Creates a dataframe with a set of potential parents for in on row for each
offspring from an Excel file.}
\usage{
getTriosDfFromExcel(triosFile)
}
\arguments{
\item{triosFile}{Character vector of length one having the path of the
trios file.}
}
\value{
A dataframe with one row for each offspring where the potential dams
and sires are the second and third column respectively. The \code{dam} and
\code{sire} columns contain the animal IDs in a single character string
separated as they were in the original file.
}
\description{
Creates a dataframe with a set of potential parents for in on row for each
offspring from an Excel file.
}
