% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTriosFileHeaders.R
\name{getTriosFileHeaders}
\alias{getTriosFileHeaders}
\title{Gets character vector of the approved file headers for a potential trios
file.}
\usage{
getTriosFileHeaders()
}
\description{
Gets character vector of the approved file headers for a potential trios
file.
}
