% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isValidAnimalAlleleHeader.R
\name{isValidAnimalAlleleHeader}
\alias{isValidAnimalAlleleHeader}
\title{Tests to see if the potential trios file has the appropriate header.}
\usage{
isValidAnimalAlleleHeader(headerLine = headerLine)
}
\arguments{
\item{headerLine}{Character vector of length one having the first line of
the file containing the list of potential trios.}
}
\description{
Tests to see if the potential trios file has the appropriate header.
}
