% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAlleleRow.R
\name{makeAlleleRow}
\alias{makeAlleleRow}
\title{Translates raw rows of an animal allele file to a data.frame row}
\usage{
makeAlleleRow(line, alleleNames, dateType)
}
\arguments{
\item{line}{A line from the animal allele file.}

\item{alleleNames}{A character vector of the header values for the alleles.
These correspond to the allele names.}

\item{dateType}{A character vector of length one used to guide
interpretation of date fields. Either "YYYYMMDD" or "mm/dd/YYYY"}
}
\value{
A vector suitable for row in an animal allele data.frame.
}
\description{
Translates raw rows of an animal allele file to a data.frame row
}
