% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFirstParentDf.R
\name{makeFirstParentDf}
\alias{makeFirstParentDf}
\title{Make first parent data.frame object for dams or sires portion of the score
object}
\usage{
makeFirstParentDf(id, firstParentType, columnNames, firstParents, scores)
}
\arguments{
\item{id}{progeny ID}

\item{firstParentType}{character vector of length one with one of \emph{dams}
or \emph{sires}.}

\item{columnNames}{character vector of columns to include from \code{scores}
in dataframe}

\item{firstParents}{character vector of parent IDs of gender firstParentType}

\item{scores}{list object with a comparison information for all trio
combinations.}
}
\description{
Make first parent data.frame object for dams or sires portion of the score
object
}
