% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeScoresDf.R
\name{makeScoresDf}
\alias{makeScoresDf}
\title{Make data.frame object from scores object}
\usage{
makeScoresDf(scores, firstParentType)
}
\arguments{
\item{scores}{list object with a comparison information for all trio
combinations.}

\item{firstParentType}{character vector of length one with one of \emph{dams}
or \emph{sires}.}
}
\value{
A data.frame object containing the following columns

\enumerate{
\item{\code{refId}} {The ID of the offspring}
\item{\code{damSex}} {The sex of the potential dam, which should always
be "F".}
\item{\code{damId}} {The ID of the potential dam}
\item{\code{damMissingLoci}} {The number of missing loci, which are the loci
without genotypes in the parent.}
\item{\code{damInvalidLoci}} {The number of invalid loci, which are the loci
missing in the offspring but present in the parent.}
\item{\code{damDiscrepantLoci}} {The number of discrepant loci, the
non-matching loci
in the parent in offspring where both have genotypes.}
\item{\code{damNumLociCompared}} {The number of loci compared}
\item{\code{damFractionNonDiscrepant}} {The fraction of non-discrepant loci,
which is
the ratio of (\code{numLociCompared} - \code{discrepantLoci}) /
\code{numLociCompared}.}
\item{\code{sire}} {A sire from the list of potential sires nested within
the current
dam of the list of dams nested inside the current offspring.}
\item{\code{sireSex}} {Sex of the sire, which should always be "M"}
\item{\code{sireMissingLoci}} {The number of missing loci, which are the
loci
without genotypes in the parent.}
\item{\code{sireInvalidLoci}} {The number of invalid loci, which are the
loci missing
in the offspring but present in the parent.}
\item{\code{sireDiscrepantLoci}} {The number of discrepant loci, the
non-matching loci
in the parent in offspring where both have genotypes.}
\item{\code{sireNumLociCompared}} {The number of loci compared}
\item{\code{sireFractionNonDiscrepant}} {The fraction of non-discrepant
loci, which is
the ratio of (\code{sireNumLociCompared} - \code{sireDiscrepantLoci}) /
\code{sireNumLociCompared}.}
}
}
\description{
Make data.frame object from scores object
}
