% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTriosRow.R
\name{makeTriosRow}
\alias{makeTriosRow}
\title{Translates raw rows of an animal trios file to a named list object}
\usage{
makeTriosRow(record)
}
\arguments{
\item{record}{A record from the trios dataframe object.}
}
\value{
A list suitable for row in an animal trios list.
}
\description{
Translates raw rows of an animal trios file to a named list object
}
