% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeAlleleRow.R, R/parseAlleles.R
\name{parseAlleles}
\alias{parseAlleles}
\title{Translates tokens of an animal allele pairs to character vectors in a list.}
\usage{
parseAlleles(tokens, alleleNames)

parseAlleles(tokens, alleleNames)
}
\arguments{
\item{tokens}{A line from the animal allele file.}

\item{alleleNames}{A character vector of the header values for the alleles.
These correspond to the allele names.}
}
\value{
A vector suitable for row in an animal allele data.frame.

A vector suitable for row in an animal allele data.frame.
}
\description{
Translates tokens of an animal allele pairs to character vectors in a list.

Translates raw rows of an animal allele file to a data.frame row
}
