% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDamsBelowMinNumber.R
\name{removeDamsBelowMinNumber}
\alias{removeDamsBelowMinNumber}
\title{Remove rows from scoresDf with dams having below minimum number compared
loci}
\usage{
removeDamsBelowMinNumber(scoresDf, minNumber)
}
\arguments{
\item{scoresDf}{dataframe with scores from \emph{scores} object}

\item{minNumber}{threshold for minimum compared loci}
}
\description{
Remove rows from scoresDf with dams having below minimum number compared
loci
}
