% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDamsOverMaxDiscrepant.R
\name{removeDamsOverMaxDiscrepant}
\alias{removeDamsOverMaxDiscrepant}
\title{Remove rows from scoresDf with dams having greater than maximum discrepant
loci}
\usage{
removeDamsOverMaxDiscrepant(scoresDf, thDiscrepant)
}
\arguments{
\item{scoresDf}{dataframe with scores from \emph{scores} object}

\item{thDiscrepant}{threshold for maximum missing loci}
}
\description{
Remove rows from scoresDf with dams having greater than maximum discrepant
loci
}
