% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeDamsOverMaxMissing.R
\name{removeDamsOverMaxMissing}
\alias{removeDamsOverMaxMissing}
\title{Remove dams with greater than maximum missing loci}
\usage{
removeDamsOverMaxMissing(scoresDf, thMissing)
}
\arguments{
\item{scoresDf}{dataframe with scores from \emph{scores} object}

\item{thMissing}{threshold for maximum missing loci}
}
\description{
Remove dams with greater than maximum missing loci
}
