% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSiresBelowMinNumber.R
\name{removeSiresBelowMinNumber}
\alias{removeSiresBelowMinNumber}
\title{Remove rows from scoresDf with sires having below minimum number compared
loci}
\usage{
removeSiresBelowMinNumber(scoresDf, minNumber)
}
\arguments{
\item{scoresDf}{dataframe with scores from \emph{scores} object}

\item{minNumber}{threshold for minimum compared loci}
}
\description{
Remove rows from scoresDf with sires having below minimum number compared
loci
}
