% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSiresOverMaxDiscrepant.R
\name{removeSiresOverMoxDiscrepant}
\alias{removeSiresOverMoxDiscrepant}
\title{Remove rows from scoresDf with sires having greater than maximum discrepant
loci}
\usage{
removeSiresOverMoxDiscrepant(scoresDf, thDiscrepant)
}
\arguments{
\item{scoresDf}{dataframe with scores from \emph{scores} object}

\item{thDiscrepant}{threshold for maximum missing loci}
}
\description{
Remove rows from scoresDf with sires having greater than maximum discrepant
loci
}
