% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/removeSiresOverMaxMissing.R
\name{removeSiresOverMoxMissing}
\alias{removeSiresOverMoxMissing}
\title{Remove sires with greater than maximum missing loci}
\usage{
removeSiresOverMoxMissing(scoresDf, thMissing)
}
\arguments{
\item{scoresDf}{dataframe with scores from \emph{scores} object}

\item{thMissing}{threshold for maximum missing loci}
}
\description{
Remove sires with greater than maximum missing loci
}
