% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{snpTrios}
\alias{snpTrios}
\title{snpTrios is a loadable version of the example snp trios data object}
\format{
An object of class \code{list} of length 7.
}
\usage{
snpTrios
}
\description{
It contains all of the trios information contained in the
"snp_potential-trios.txt" file.
}
\examples{
\donttest{
library(parentfindr)
data("snpTrios")
head(trios)
}
}
\keyword{datasets}
