% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sortScores.R
\name{sortScores}
\alias{sortScores}
\title{Sorts the \code{scores} object}
\usage{
sortScores(scores)
}
\arguments{
\item{scores}{A list object that can contain the scores of a single potential
parent a potential mates or scores of multiple potential parents and their
potential mates.}
}
\value{
A sorted \code{scores} object

The \code{scores} object can contain the scores of a single potential parent
a potential mates or scores of multiple potential parents and their
potential mates.
}
\description{
Sorts the \code{scores} object
}
