% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stripComments.R
\name{stripComments}
\alias{stripComments}
\title{Strip off comments}
\usage{
stripComments(stmt)
}
\arguments{
\item{stmt}{character vector containing XML text}
}
\value{
Character vector with all XML comments removed.
}
\description{
Takes a character vector containing XML text and removes
comments. Lines that contained only comments are also removed.
}
