% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addSecondParents.R
\name{addSecondParents}
\alias{addSecondParents}
\title{Add second parent data to first parent data.frame object.}
\usage{
addSecondParents(id, firstParentDf, secondParentType, columnNames, scores)
}
\arguments{
\item{id}{offspring ID}

\item{firstParentDf}{first parent data.frame object for dams or sires
portion of the score object dependent on which is the first parent.}

\item{secondParentType}{character vector of length one with one of \emph{dams}
or \emph{sires}.}

\item{columnNames}{character vector of columns to include from \code{scores}
in dataframe}

\item{scores}{list object with a comparison information for all trio
combinations.}
}
\description{
Add second parent data to first parent data.frame object.
}
