% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScores.R
\name{computeScores}
\alias{computeScores}
\title{Scores a potential parent to a kid and optionally a previously scored parent.}
\usage{
computeScores(parent, kid, other)
}
\arguments{
\item{parent}{A list with the all of the information from the animal allele
file representing one potential parent.}

\item{kid}{A list with the all of the information from the animal allele
file representing the potential offspring.}

\item{other}{A list with the all of the information from the animal allele
file representing one potential parent.}
}
\value{
A list containing the offspring \code{refId}, the \code{sex} of the
potential parent that was scored, the \code{pkMatch} list, which contains
offspring's \code{refId} and the two logical vectors of left and right
allele comparison results between the potential parent and the offspring,
and finally it has the characer vector of text, which fully describes the
discrepancies with one discrepancy per element.
}
\description{
Scores a potential parent to a kid and optionally a previously scored parent.
}
