% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computeScores.R
\name{computeScores}
\alias{computeScores}
\title{Counts the number of loci with missing loci data}
\usage{
computeScores(parent, kid, other)
}
\arguments{
\item{parent}{A list with the all of the information from the animal allele
file representing one potential parent.}

\item{kid}{A list with the all of the information from the animal allele
file representing the potential offspring.}

\item{other}{A list with the all of the information from the animal allele
file representing one potential parent.}
}
\value{
An integer value equal to the number of loci missing loci data
}
\description{
Counts the number of loci with missing loci data
}
