% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFirstParentScores.R
\name{getFirstParentScores}
\alias{getFirstParentScores}
\title{Get first parent score from scores object}
\usage{
getFirstParentScores(id, firstParent, firstParentType, columnNames, scores)
}
\arguments{
\item{id}{progeny ID}

\item{firstParent}{first parent ID}

\item{firstParentType}{character vector of length one with one of \emph{dams}
or \emph{sires}.}

\item{columnNames}{character vector of columns to include from \code{scores}
in dataframe}

\item{scores}{list object with a comparison information for all trio
combinations.}
}
\description{
Get first parent score from scores object
}
