% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrios.R
\name{getTrios}
\alias{getTrios}
\title{Tests to see if the potential trios file has the appropriate header.}
\usage{
getTrios(lines)
}
\arguments{
\item{lines}{Character vector of length one plus the number of animals where
the additional line is the header}
}
\description{
Tests to see if the potential trios file has the appropriate header.
}
