% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTrios.R
\name{getTrios}
\alias{getTrios}
\title{Detect file type of trios file and routes its parsing.}
\usage{
getTrios(triosFile)
}
\arguments{
\item{triosFile}{Character vector of length one having the path of the
trios file.}
}
\description{
The trios files are made up of sets of potential parents in one row for each
offspring from CSV file.
}
