% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeTriosRow.R
\name{makeTriosRow}
\alias{makeTriosRow}
\title{Translates raw rows of an animal trios file to a named list object}
\usage{
makeTriosRow(line)
}
\arguments{
\item{line}{A line from the animal trios file.}
}
\value{
A list suitable for row in an animal trios list.
}
\description{
Translates raw rows of an animal trios file to a named list object
}
